Reading of MCD43 data and creation of ObsOp
The code is able to read MCD43 datasets, and create the
observation operator using an emulator. There are some details on
how this is done, but it is broadly working. It might be too heavy
to do on entire MODIS tiles, however, but this is yet to be tested.
Currently, using a 512*512 image chip.