Dealing with multiband datasets
This commit makes the start of using multiband datasets. As KaFKA
is designed to exploit the linearisations in the model, we can in
effect assimilate one band with one particular linearisation,
update the state with a new band and so on. But that means that
the state after a few bands could be quite different. We can thus
now iterate over all the bands again updating the observation
operator for each of them. This is now implemented, but as I
haven't really used a multiband dataset, mostly untested.