Splitting solvers from main class
**This is WiP**

The first tests show that solving the KF update on large state
vectors is very expensive in terms of memory. A better approach
might be to just use a slightly different approach (a so-called
variational Kalman filter), but this neesd that the LinearKalman
class has the assimilation bit spun out of the class. I have
started this and put the method in `solvers.py`, but I haven't
yet tested. This should also make testing easier, as you can just
test the solver without the filter. The idea is that the user
selects the solver in the code.