A full raster observation operator
These changes are moving to a full raster observation operator.
The first thing to do is to provide H for all locations. If the
locations are masked, then we just provide 0. In general, this
leads to not being able to solve for the state in the standard
KF, but the version I'm using now is always supported by the prior.