Variational Kalman solver
The method is supposeed to work with a full state representation,
where locations that have no data will just be the prior. The
solver is quite fast for this user case at any rate. The other
important change is that I assume that the state has the parameters
for each pixel sequentially (I don't think this matters much here,
but it makes the inverse covariance block diagonal, which is nice)