Dealing with sparse mtcs and other stuff
The main changes are to do with the specification of the dynamic
model. It now needs to specify the size of the state grid, which
just means a couple of extra parameters when calling the methods.

The propagation of the state is a bit different if only the
inverse analysis covariance is available. Basically, I just
inflate the main diagonal. Bit of jiggery pokery...