Working version, needs better plotting
This last update has provided the following important things:

* The variational solver has been updated. Previously, the
linearised code was wrong. I have now updated this, and things
appear to be going in the right direction.
* The model now moves from one date to the next fine.
* Iterations for a particular date work OK (I think...), and the
data are read only once at the beginning of the loop, so iterations
are faster (a bit more memory gets used...).

Currently, the biggest bottleneck is populating the H matrix. Using
the emulators is expensive, but re-arranging the output is even more
expensive.