Testing
The nonlinear code is now undergoing testing. This means that I
now have an "state" advance code that makes the uncertainty (and
the state) rever to Pinty's prior. In effect, this is a linear
version of the JRC-TIP. The plots have been slightly improved by
putting parameter boundaries, and transformed LAI has been
acknoweledged throughout the code.