[WiP] Refactoring things a bit
The main idea behind this branch is to allow the user to run the
the filter on demand when a new observation arrives. This requires
that the main loop in the `run` method in `linear_kf.py` is split
into a method that allows updating. As a result, the `run` method
should also be rafactored. I have also added an observations class
which acts as a storage for data. The aim is that this class is the
one that the Kalman filter interfaces with to acquire data, masks,
uncertainty and (possibly) emulators.

The idea is that the observations class gets updated with new
observations mapped by time, and the KF is able to request the
observations from a particular date.