Cleaning up
I have removed the diagonal solver as we don't really use it
anymore. The class has a method to create observation operators
that allows the user to change for a linear or a non-linear model,
and the observations are also read in by calling an external
function. Also, time steps and everything are now given as
datetime objects, which ends up being slightly cleaner.