Potentially limiting the number of emulator runs
Main bottlenecks right now are:
1. Evaluating the emulators for lots of pixels
2. Putting the results of the emulation into a matrix.

The second part is usual Python loop, and could be alleaviated using
numba. The first is a bit more complicated. I have tried a number
of things. One is to just run once if all pixels are the same, and
I then extended this concept with some fancy clustering (so you
do say 500 emulator runs). It turns out that the clustering idea
is really stupid, as clustering takes longer than running the
emulators D'Oh!!! It would be far better to just do a LHS design
over parameter space, or even just regular sampling and nearest
neighbour attribution. In this commit, the clustering is switched
off, but the code is there.