Noddy LUT implementation
Rather than waste time with clustering and whatnot, let's have a
noddy LUT. I have added a function to find the closest element to
a vector in a set of vectors of the same length. The function is
very simple, and I'm putting some comments here in case I forget about
how it works....

Let's say that we have a state vector with 10000*7 parameters, which we
can recast as a (im) [10000,7] array. From here we select (e.g.) 500
samples that span this space, and we end up with a LUT (lut) of [500,7].
We also run the LUT through the emulator to get H_ and dH_ (500 vector
and [500,7] gradient matrix). We want to then assign the original
parameters to individual H_ and dH_ elements based on closes distance
between parameters. We can get for each pixel the smallest distance with
some clever array broadcasting:

np.linalg.norm(lut[:, None, :] - im, axis=2).argmin(axis=0)

The difference calculates the difference between each element of the LUT
and each element of the array for each dimension. We then calculate the
l2 norm along the last axis (which in this case is the 7 dimension), and
we end up with a [150, 10000] array, which contains the euclidean
distance from each element in the state to each element in the LUT. The
argmin bit on the first axis then results in a 10000 array indicating
the location in the LUT that is closest to the each vector in the state.
Then there's no need to use any extra library!