LUT emulator speed up -- Working (ish)--
This version implements an approximation to the creation of the
forward model through a clustering of the input parameter space.
Basically,

1. We check the number of samples,
2. and if it is too big for convenient emulation...
3. We fit a multivariate Gaussian to the state samples, and draw eg 1500
samples
4. We assign the closest point from (3) to each sample in the input
state vector
5. We run the emulator for the 1500 samples, and store the output and
gradient
6. We assign the output and gradients to the full vector using the
assignation mapping from point (4).

It seems to work OK, and it's fast, but I'm not sure whether this
is a safe thing to do, or we need a slightly different strategy for
this. And it needs testing.