Updates to state temporal grid
The previous temporal grid checked for all available observations
at the time of each grid point, which is fine if you're working
on daily data. The new version checks for the observations between
time points. It needs checking, but I first need to push to gh to
test on UCL's system :-/