Added a BHR test class
The new observation class allows one to hopefully read in a single
"pixel" of BHR data over time to do some quick tests. In essence,
it just needs a set of dates, albedos and a gp (this bit is missing)
It then needs to pack the data into a BHR structure.