Towards multiband assimilation
This commit (and branch) aim to invert observations by taking all
bands simultaneously. Theory for this is that observations for all
bands are stacked, and so are the H_matrices.

There's a new `assimilate_multiple_bands` method that loops over all
the observations.

The `do_all_bands` method is the main iteration for the solution,
which broadly reads in the data for all bands, and then linearises
the observation operator for all available bands. Although it's a
bit unclean, I have tried to keep the same interface to call `solver`
directly, although I called it `solver_multiband` for the time being.

`solver_multiband` calls a version of `variational_kalman`, where
the new matrices and vectors will be put together. Basically, it's only
in calculating the (H(x)-y).TC_obs(H(x)-y) that the changes are
relevant: H maps from `self.n_params*n_pxls` to `n_bands*n_pxls`. So
all this needs is careful stacking of observations. I haven't done
this yet...

@GerardoLopez can you take a look please?