Changes to work with multiple bands
I have now tested things. This version produces something akin to
a reasonable solution. Please test, and
**NOTE** the Hessian corrections are still missing. I think one
ought to be able to calculate the correction for individual bands
and add them together, but this requires having access to the
innovations from both bands (as well as uncertainties for both
bands too).