Robust strategy
Not quite there on this one. The robust function is here, but I
think this needs to be applied only once after the first iteration,
as otherwise, you get bogged down in endless moving goalpoasts in
what is an outlier.

Also, one could argue that the first iteration might be a pretty
rubbish prediction anyway, so better to let the solver go for a
couple of iterations before going through this. Jury still out as
to what is the best approach.

Additionally, one has to think about what to do depending on the
distance between the outlier and the prediction. A first pass might
be to say that if the z score is > 5(?) sigma, you just update the
mask. The other more subtle option would increase the observational
uncertainty in the observations if the 3 <=  z_score < 5 (say),
mimicking a Huber norm.