Dynamic model is now in its own class
The dynamic model that advances the state between time steps is
now whisked off to a specialised class that gives a lot more
flexibility. This requires a number of changes to provide the
linear model transformation and inflation matrices at each time
step, but all it does is really clean the code up a bit.