Smarter convergence
I have noticed that convergence sometimes falls off as iterations
mount up. So I have added a flag that switches further iterations
if the convergence metric increases. This is a blunt tool, as I
don't really know why these increases happen. Part of it might be
due to outliers, or it might just be due to other horrors lurking
beneath the surface.