Can now run forwards and backwards
The filter can now be run forwards and backwards. This is done
by setting the `time_grid` to be monotonically increasing (as it
was done up to now), or monotonically decreasing (e.g. start at the
end of the temporal window and go backwards towards the start).

This allows one to build up a smoother.